+++ packages/mpack/src/utils/printErrors.ts
@@ -1,0 +1,5 @@
+export function printErrors(errors: PromiseRejectedResult[]) {
+  for (const error of errors) {
+    console.error(error.reason);
+  }
+}