+++ packages/plugin-core/src/index.ts
@@ -5,6 +5,7 @@ export * from './utils/mergeConfig';
 export * from './utils/mergeBuildConfigs';
 export * from './utils/createPluginHooksDriver';
 export * from './utils/buildResult';
+export * from './utils/resolveConfig';
 
 export { defineConfig } from './config/defineConfig';
 export { loadConfig } from './config/loadConfig';