+++ infra/forge-cli/CHANGELOG.md
@@ -1,5 +1,13 @@
 # @granite-js/forge-cli
 
+## 0.1.17
+
+### Patch Changes
+
+- Updated dependencies [9c415df]
+  - @granite-js/plugin-core@0.1.17
+  - @granite-js/deployment-manager@0.1.17
+
 ## 0.1.16
 
 ### Patch Changes